+++ napari/_qt/qt_eliding_label.py
@@ -1,0 +1,53 @@
+from qtpy.QtCore import QPoint, QRect, QSize, Qt
+from qtpy.QtGui import QFontMetrics, QPainter, QTextLayout
+from qtpy.QtWidgets import QFrame
+
+
+class MultilineElidedLabel(QFrame):
+    def __init__(self, parent=None, text=''):
+        super().__init__(parent)
+        self.setText(text)
+
+    def setText(self, text=None):
+        if text is not None:
+            self._text = text
+        self.update()
+        self.adjustSize()
+
+    def paintEvent(self, event):
+        super().paintEvent(event)
+        painter = QPainter(self)
+        font_metrics = painter.fontMetrics()
+        text_layout = QTextLayout(self._text, painter.font())
+        text_layout.beginLayout()
+
+        y = 0
+        while True:
+            line = text_layout.createLine()
+            if not line.isValid():
+                break
+            line.setLineWidth(self.width())
+            nextLineY = y + font_metrics.lineSpacing()
+            if self.height() >= nextLineY + font_metrics.lineSpacing():
+                line.draw(painter, QPoint(0, y))
+                y = nextLineY
+            else:
+                lastLine = self._text[line.textStart() :]
+                elidedLastLine = font_metrics.elidedText(
+                    lastLine, Qt.ElideRight, self.width()
+                )
+                painter.drawText(
+                    QPoint(0, y + font_metrics.ascent()), elidedLastLine
+                )
+                line = text_layout.createLine()
+                break
+        text_layout.endLayout()
+
+    def sizeHint(self):
+        font_metrics = QFontMetrics(self.font())
+        r = font_metrics.boundingRect(
+            QRect(QPoint(0, 0), self.size()),
+            Qt.TextWordWrap | Qt.ElideRight,
+            self._text,
+        )
+        return QSize(self.width(), r.height())
